Do not fail on qbootctl -m if device is not A/B
It is common and makes sense to ship qbootctl -m as some kind of init
service that runs at boot. However, not all devices are A/B partitioned
and qbootctl -m will currently return an exit code of 1 on these
devices, making the init service marked as failed. We should handle this
case more gracefully and exit with exit code 0 if there was nothing to
do.

Signed-off-by: Jens Reidel <adrian@travitia.xyz>